requirements.txt
asn1crypto==1.5.1
autocommand==2.2.2
bcrypt==5.0.0
blinker==1.9.0
Brlapi==0.8.8
certifi==2026.2.25
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
cupshelpers==1.0
dasbus==1.7
dbus-python==1.4.0
distro==1.9.0
Flask==3.1.3
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
fuse-python==1.0.9
greenlet==3.3.2
html5lib-modern==1.2
httplib2==0.22.0
idna==3.11
inflect==7.5.0
itsdangerous==2.2.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
kali-tweaks==2025.4.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lightdm-gtk-greeter-settings==1.2.3
louis==3.37.0
MarkupSafe==3.0.3
mechanize==0.4.10
more-itertools==10.8.0
netifaces==0.11.0
oauthlib==3.3.1
onboard==1.4.3.post9
packaging==25.0
pexpect==4.9.0
psutil==7.1.0
ptyprocess==0.7.0
pycairo==1.27.0
pycups==2.0.4
PyGObject==3.54.5
PyJWT==2.11.0
pylibacl==0.7.2
pyparsing==3.1.3
PyQt6==6.10.2
PyQt6_sip==13.11.0
pysmbc==1.0.25.1
python-apt==3.1.0
python-dotenv==1.2.2
pyxattr==0.8.1
pyxdg==0.28
requests==2.32.5
sentry-sdk==2.22.0
setproctitle==1.3.7
SQLAlchemy==2.0.48
tornado==6.5.4
typeguard==4.4.4
typing_extensions==4.15.0
urllib3==2.6.3
wadllib==2.0.0
webencodings==0.5.1
Werkzeug==3.1.6
wheel==0.46.3
xdg==5
